// src/openapi/model/user.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Role } from './role';


export interface User { 
    birthDate?: string;
    createdDate?: string;
    email?: string;
    firstName?: string;
    id?: number;
    lastModifiedBy?: string;
    lastModifiedDate?: string;
    lastName?: string;
    password?: string;
    phoneNumber?: string;
    recordStatus?: number;
    role?: Role;
    username?: string;
}